fixup! fixup! fixup! fixup! Added transient caching for the product listing variation sliders for improving performance.
