+++ src/WooCommerce.php
@@ -13,6 +13,13 @@ namespace Netzstrategen\Gallerya;
 class WooCommerce {
 
   /**
+   * The slider variation transient key prefix.
+   *
+   * @var string
+   */
+  const VARIATION_SLIDER_CACHE_KEY_PREFIX = Plugin::PREFIX . '_variation_attachments_for_';
+
+  /**
    * Adds woocommerce specific settings.
    *
    * @implements woocommerce_get_settings_gallerya
@@ -102,7 +109,11 @@ class WooCommerce {
 
     if ($product->is_type('variable')) {
       // Checks for cached content, avoiding too many queries.
-      $attachment_ids = Cache::getCachedVariationSliderAttachmentIds($product);
+      $attachment_ids = get_transient(self::VARIATION_SLIDER_CACHE_KEY_PREFIX . $product->get_id());
+      if (empty($attachment_ids)) {
+        // Builds cache data, so next time it will load faster.
+        $attachment_ids = self::setProductVariationTransients($product);
+      }
 
       // Only render slider if there are more than one images.
       if (count($attachment_ids) > 1) {
@@ -120,6 +131,32 @@ class WooCommerce {
       // output default thumbnail markup.
       woocommerce_template_loop_product_thumbnail();
     }
+  }
+
+  /**
+   * Builds transient record for this variation slider.
+   */
+  private static function setProductVariationTransients($product) {
+    $attachment_ids = [];
+    // Gets the main product image.
+    $attachment_ids[] = $product->get_image_id();
+
+    // Gets the first image of each product variation.
+    $variations = $product->get_available_variations();
+    foreach ($variations as $variation) {
+      $attachment_ids[] = $variation['image_id'];
+    }
+    $attachment_ids = array_unique($attachment_ids);
+
+    set_transient(self::VARIATION_SLIDER_CACHE_KEY_PREFIX . $product->get_id(), $attachment_ids);
+    return $attachment_ids;
+  }
+
+  /**
+   * Deletes variation slider transient certain product id.
+   */
+  public static function flushVariationAttachmentsTransients($product_id) {
+    delete_transient(self::VARIATION_SLIDER_CACHE_KEY_PREFIX . $product_id);
   }
 
 }
